requirements.txt
absl-py==0.11.0
albumentations==0.5.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
efficientnet==1.0.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
image-classifiers==1.0.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.19.4
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
opt-einsum==3.3.0
packaging==20.8
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
scikit-image==0.17.2
scipy==1.4.1
segmentation-models==1.0.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.9.3
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
